fix(motionplayer): M2 — EmoteAngleController powCount is raw float bits (completes float-bits sweep)
EmoteAngleController_step@0x666634 reads pow(v11, *(float*)(a1+100)) at 0x6666f8
(LDR S, no SCVTF) and writes a1+100 as a raw DWORD copy of keyframe+8
(*(a1+100) = *(uint*)(v6+8) at 0x6667e4). The port stored powCount as int32_t
and used static_cast<float> — same raw-bits-vs-int bug as the other controllers.

Fixed: EmoteAngleController::powCount int32_t -> float, populated via memcpy from
the keyframe's uint32 powCount, used directly in std::pow. This completes the
float-bits correction sweep across all EmoteEngine controllers (eye/eyebrow/
mouth in 2316276/m2mouth, transition/var + selector in 2870209, angle here).

Build: web/debug exit 0, grep error: = 0.

Not runtime-verified: the angle controller's curved-rotation interpolation has no
oracle/fixture; verification = decompile (0x666634) + build.

+++ cpp/plugins/motionplayer/EmoteAngleController.h
@@ -10,7 +10,7 @@
 //   +88       float    targetRad     // after shortest-path wrap
 //   +92       float    startRad
 //   +96       float    invDuration
-//   +100      int32_t  powCount
+//   +100      float    powCount (raw float bits from keyframe[+8], read *(float*))
 //   +104      float    phase
 //   +108      int32_t  pad
 //
@@ -48,7 +48,7 @@ namespace motion {
         float   targetRad  = 0.0f;  // +88
         float   startRad   = 0.0f;  // +92
         float   invDuration = 0.0f; // +96
-        int32_t powCount = 0;       // +100
+        float   powCount = 0.0f;    // +100 — raw float bits (read *(float*), not int)
         float   phase = 0.0f;       // +104
         int32_t pad = 0;            // +108
     };
